Fix float32 precision loss for large timestamps in plotting
Store trace timestamps in Float64Array instead of Float32Array to
preserve precision. In the rendering pipeline, subtract a base time
offset before storing x-coordinates in the GPU's Float32 buffers,
and compensate by shifting the viewport accordingly per line.

+++ js/traces.js
@@ -44,7 +44,7 @@ class SeriesDataChunk {
         this.name = name;
         this.size = size;
         this.dim = dim;
-        this.times = new Float32Array(size);
+        this.times = new Float64Array(size);
         this.timedData = Array.from(Array(dim)).map(_ => new Float32Array(size));
         this.timedDataYLimits = Array.from(Array(dim)).map(_ => new Lim(0, 0));
         this.timeFrom = 0;
